requirements.txt
SpeechRecognition==3.8.1
numpy==1.24.4
matplotlib
easygui
Pillow

# SpeechRecognition==3.8.1
# numpy==1.18.2
# opencv-python==4.2.0.34

